refactor: split backend logic into lib modules
